Added workflow segmentation algorithm training.
git-svn-id: https://subversion.assembla.com/svn/slicerigt/trunk/Slicer4/WorkflowSegmentation@196 35aa5c59-fbbe-4c73-a865-bce86c728669

+++ Logic/RecordType.h
@@ -3,6 +3,7 @@
 #define RECORDTYPE_H
 
 #include <vector>
+#include <sstream>
 
 class ValueRecord
 {
@@ -29,7 +30,7 @@ class TimeLabelRecord : public LabelRecord
 {
 public:
   double time;
-  void setTime( int newTime ) { time = newTime; };
+  void setTime( double newTime ) { time = newTime; };
   double getTime() { return time; };
 };
 
@@ -44,5 +45,6 @@ public:
   int getState() { return state; };
   int getSymbol() { return symbol; };
 };
+
 
 #endif
